+++ multibody/fem/mpm-dev/BUILD.bazel
@@ -18,7 +18,9 @@ drake_cc_binary(
         "main.cpp",
     ],
     deps = [
+        ":CorotatedModel",
         ":MPMDriver",
+        ":SaintVenantKirchhoffWithHenckyModel",
         "//common:essential",
         "//math:geometric_transform",
         "@partio",
@@ -60,6 +62,36 @@ drake_cc_library(
         "ConstitutiveModel.h",
     ],
     deps = [
+        "//common:essential",
+        "//multibody/fem:matrix_utilities",
+    ],
+)
+
+drake_cc_library(
+    name = "CorotatedModel",
+    srcs = [
+        "CorotatedModel.cc",
+    ],
+    hdrs = [
+        "CorotatedModel.h",
+    ],
+    deps = [
+        ":ConstitutiveModel",
+        "//common:essential",
+        "//multibody/fem:matrix_utilities",
+    ],
+)
+
+drake_cc_library(
+    name = "SaintVenantKirchhoffWithHenckyModel",
+    srcs = [
+        "SaintVenantKirchhoffWithHenckyModel.cc",
+    ],
+    hdrs = [
+        "SaintVenantKirchhoffWithHenckyModel.h",
+    ],
+    deps = [
+        ":ConstitutiveModel",
         "//common:essential",
         "//multibody/fem:matrix_utilities",
     ],
@@ -270,6 +302,7 @@ drake_cc_googletest(
 drake_cc_googletest(
     name = "Particles_test",
     deps = [
+        ":CorotatedModel",
         ":Particles",
         ":TotalMassAndMomentum",
         "//common/test_utilities:eigen_matrix_compare",
@@ -293,6 +326,8 @@ drake_cc_googletest(
     name = "ConstitutiveModel_test",
     deps = [
         ":ConstitutiveModel",
+        ":CorotatedModel",
+        ":SaintVenantKirchhoffWithHenckyModel",
         "//common:essential",
         "//common/test_utilities:eigen_matrix_compare",
         "//math:geometric_transform",
@@ -332,6 +367,7 @@ drake_cc_googletest(
     name = "MPMTransfer_test",
     deps = [
         ":BSpline",
+        ":CorotatedModel",
         ":Grid",
         ":MPMTransfer",
         ":Particles",
