+++ multibody/fem/mpm-dev/test/Particles_test.cc
@@ -9,6 +9,7 @@
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/math/rotation_matrix.h"
 #include "drake/multibody/fem/mpm-dev/ConstitutiveModel.h"
+#include "drake/multibody/fem/mpm-dev/CorotatedModel.h"
 #include "drake/multibody/fem/mpm-dev/TotalMassAndMomentum.h"
 
 namespace drake {
